The classfile writer now handles the creation of interfaces.
JVM::MAKE-CLASS-INTERFACE-FILE will create a classfile structure
suitable for defining a Java interface.  Simply add abstract methods
to complete the definition.

JVM::ADD-SUPERINTERFACE adds an interface definition to a classfile.
For a classfile denoting a Java object, these interfaces correspond
to the contents of a Java language "implements" clause.  For a classfile
denoting a Java interface, these interfaces are all added to the
contract defined by this interfaces corresponding to the Java language
"extends" clause.

[Includes accidental update of NetBeans project files.]