Fix ad infinitum appending of 'svnversion' results in incremental builds.
The 'version' file is now removed during each invocation of
'abcl.stamp.version' which adds a bit of time to each build, but at
least implement the correct semantics (i.e. always produce an accurate
svnversion result).  An optimization might be to conditionally compare
the contents of 'version' with the results of running a fresh
'svnversion', but to do this with Ant's tiny little insect brain
requires more effort than this mammal deems wise to exert.