Provisionally working version of JSS without bsh-2.0b4.jar.
This still needs 'jscheme.jar' to be loaded via the top-level
declaration at the beginning of packages.lisp.  Adjust the filepath to
a local version of jscheme.jar which may be downloaded from
http://code.google.com/p/lsw2/source/browse/trunk/lib/jscheme.jar.

Rigourously untested, but still a worthwhile checkpoint for public
consumption, especially since we need to fix on an API.

Re-packaged in JSS package.  Use ENSURE-COMPATIBILITY to be compatible
with existing JSS installations.