Fix JAR-PATHNAME tests.
Remove the wrong-headed use of TRUENAME at compile time which would
error unless the test directory had previously been created.

Remove the use of the readtime #. macro to create the success
conditions for tests which were failing based on the presence of the
temporary jar directory.

These bugs seem show the need for a more explicit test
build-up/tear-down sequence as the complexity in manage Lisp macros
here leads to a lot of unncessary bugs.