weblogic:  ABCL loads under the Weblogic 10.3 application server.
If the Pathname parsing try to parse a URI but get an empty authority,
don't push the values to the HOST element.  Addtionally warn if we
parse a null authority via the Debug.trace() facility.

Weblogic uses the "zip" scheme which more or less maps to ABCL's
nesting of "file" within "jar".  This patch doesn't systematically
follow the logical consequences of that disinction, but it does allow
ABCL to load its own fasls under weblogic-10.3.  The right thing will
be to establish some sort of ability to run the ABCL tests inside of
various application server containers, but that will take a fair amount of resources.