Fix UNEXPORT to work on symbols from foreign packages.
No longer check that the symbols which are the target of UNEXPORT are
accessible.  Such symbols may be present in a foreign package as they
may have been part of a USE clause for which the original symbol has
subsequently made internal in its home package by a previous UNEXPORT
operation.

Fixes #311.