{decode,encode}-universal-time "time of the time" timezone semantics
Scott L. Burson

The implementation of EXT:GET-TIME-ZONE provides the underlying
mechanism.

Both GET-TIME-ZONE and DEFAULT-TIME-ZONE have been elevated to
supported status by virtue of promotion of their respective symbols to
the EXTENSIONS and SYSTEM packages, respectively.

ME: A mechanism for adding symbols to both SYS and EXT packages with
DEFINE-SYMBOL-MACRO didn't work, but left as comment at the end of
'time.lisp' for further work.