Define undefined conditions and handler functions following error-fun='error pattern (Alan Ruttenberg)
Activation of function are in in comments for now, pending solving boot issues.


On malformed destructuring-bind, bogus-sublist-error was being
signaled, but said condition didn't exist, and so true error was
masked.

On review this code seems to be adapted from CMUCL, but special casing
error-fun = 'error and using program-error instead of the conditions.

I brought over the conditions from CMUCL, which are in a comment at
the bottom - they should loaded only after clos - wasn't sure of the
mechanics.

I then defined a function for bogus-sublist error, based on the
reports for the condition, and replaced arg-count-error with the more
detailed report as well.

Merges <https://github.com/armedbear/abcl/pull/61>.

From
<https://github.com/armedbear/abcl/pull/61/commits/bdf75f5adeda85ca0a587b51c474112b825b5659>,
<https://github.com/armedbear/abcl/pull/61/commits/78c7e0e3291ee97fbac2a8fa26bc55cf180b0b41>.