Add jtypecase form, fix print-object of null pointer (Alan Ruttenberg)
Like typecase except the check is whether the form is an instance of
one of the classes, which are lookup up using find-java-class.

Also included: a minor fix to print-java-object-by-class - it was
getting an error while printing a null pointer.

The printer now prints null references as "#<null>".

Merges <https://github.com/armedbear/abcl/pull/73>