Consolidated patches for running under Java 11
The ABCL-CONTRIB mechanism now works, and JSS has been patched to work
as well.  Surprisingly, everything seems work under both openjdk-8 and
openjdk-11 without the major modifications that were feared to be
necessary, including the substantial monkeying around with the
accessibilty conventions which JSS works through.

This patch has been mostly tested under openjdk-11 and openjdk-8 for
macOS/Linux.

We rework ABCL-CONTRIB strategy to introspect the value of the the
Java system property "java.class.path".  This may not be sufficient to
find the contribs under all possible ways that the implementation may
be hosted, i.e. in various cloud based hosting environments, but
seemingly works fine from the command line.

The changes to abcl-contrib and JSS are somewhat rough in that while
they work, they could use a refactoring to be a bit less hacky
looking, as well as name things consistently.