abcl-asdf: rework Maven usage strategy
On initialization of the Maven Aether linkages, If
ABCL-ASDF::*MVN-LIBS-DIRECTORY* is set, use that instead of finding
the path by invoking 'mvn'.

Don't depend on the invocation of `mvn` to determine things as it is
both slow, and error prone if JAVA_HOME is not set correctly.  We have
used the libraries themselves rather than invoking the
executable.

Since by definition, ABCL-ASDF depends on ASDF/UIOP being present, use
UIOP:RUN-PROGRAM instead of SYS:RUN-PROGRAM as it seems to work better
under Java 6.

Re-worked IGNORES-ERROR invocation order so that we ar now working
with maven-3.0.4 and maven-3.6.3.

Export MVN-VERSION.

Document the current compatibility status somewhat.

TODO: provide a restart that will download locally a version
of Maven if none can be found.

TODO: chase down the compile-time binding of
MavenRepositorySystemUtils in the fasl which can cause load failure
when switching Maven versions used by the local implementation.