abcl-asdf: use WITH-AETHER to ensure things are loaded
Unfortunately, since we can't easily "unload" jar files from the
classpath, it isn't easy to make the WITH-AETHER macro work with
differing versions of Aether.  Still, it's a little cleaner to use a
WITH-AETHER form to ensure that the Maven initialization has been done
correctly.

TODO: ensure that WITH-AETHER invoked with a bogus Maven Aether
library location doesn't bork the system.