ci: just output hashes for release
Installing the necessary TeXlive chain to render the manual proves to
be a very time-wise expensive operation on Travis CI for a generic
container.  For now, disable the creation of full release artifacts
just printing the hash and sizes for what exists.

Use and set the environment variable ABCL_ROOT to start to make the CI
instructions implementation independent.