Re-work strategy for CL:*FEATURES* for hosting JVM
Add a keyword denoting the Java platform to CL:*FEATURES* of the form
"JAVA-<n>" where "<n>" is an integer denoting the major version of the
platform according to JEP-223.

We deprecate depending on the presence of JAVA-1.6, JAVA-1.7, and
JAVA-1.8 in favor of JAVA-6, JAVA-7, and JAVA-8.

Add the contents of the "os.arch" property if not running under x86.

c.f. <https://openjdk.java.net/jeps/223>

TODO:  describe the presence of of various CL:*FEATURES* in the Manual.