INCOMPLETE pathname: overhaul semantics for zips using PATHNAME-DEVICE
Implement CL:PATHNAME-DEVICE references to files which are cons cells
to the level of abcl-1.7.0, which is that nested pathnames do not
fully work for all variants of CL:TRUENAME.

Attempt to start coding "worst case, cache everything we find
strategy".

TODO memoize storage via a weak reference so that the VM can possibly
reclaim space if it needs to.

TODO re-implement references to URIs