INCOMPLETE Implement Zip from InputStream access
ZipCache.Archive.getEntry() now works on pathnames relative to the
root of the Archive.

Able to retrieve the reference to the InputStream for a nested pathname.

Now need to implement accessing entries within this reference.

TODO Rework decision to always use fully qualified pathname in the cache.
* * *
Fix Pathname.copyFrom() not to change order of DEVICE cons

Reproduce via:

   #p"jar:jar:file:/var/tmp/cl-ppcre-all-2.1.1.jar!/cl-ppcre/packages.abcl!/**/*"
   (namestring *)
* * *
PROBE-FILE fails for archive stream

Somewhere in the code we can construct root jar as a Pathname rather
than PathnameURL which means that the two values are not
java.lang.Object.equals(), so not matched.

TODO figure out how we are getting such a value
* * *
Start being more explicit about the truename method
* * *
Delegate cloning of device to the JAR-PATHNAME Java object

TODO PathnameXXX.createXXX() routines uniformly return object over
LispObject.  Is this really necessary?  It seems to work now in a kind
of "mixed" state.