asdf: fix PARSE-UNIX-NAMESTRING in presence of JAR-PATHNAME device
Under ABCL, when loading system definitions from a JAR-PATHNAME, the
PARSE-UNIX-NAMESTRING was including the DEVICE in the defaults to
MAKE-PATHNAME.  This was mostly harmless but fails when the as yet
unreleased abcl-1.8.1 tightened the definition of JAR-PATHNAME to
always have an absolute directory.

TODO: ensure that this works correctly in the presence of logical
pathnames, and under Windows where the device may contain the logical
drive letter or the UNC path.