build: deprecate use of abcl.implementation.version
Use abcl.version instead, fixing manifests for jar packaging.

The usage of abcl.implementation.version is now deprecated as not
working for a long time.  The intention was that if a
given source tree could be unidentified as being under source control
by the presence of given revision control system artifacts, the
abcl.implementation.version would contain a version derived from the
revision control system.

To hack around the non-prescriptive nature of Ant, chains of
conditional targets had to be created that while might have once
worked for Subversion, were going to get increasingly hairy for
Mercurial and Git.