Correct ENCODE-UNIVERSAL-TIME for dates after 2037
The code removed in this patch seems to be attempting to adjust for
times which occurred after a 32-bit representation of the UNIX epoch
would wrap around to zero.  This doesn't make much sense as Common
Lisp times are clearly defined as an integer number of seconds after
1900.

The ansi-test DECODE-UNIVERSAL-TIME.5 now passes.

c.f. <https://abcl.org/trac/ticket/443>.