tools: start working towards a documentation of code style
We wish commits in general to be "non-noisy" with respect to
whitespace changes, but the codebase is not quite at a common baseline
due in part to the lack of a good AST manipulation tool for Java, but
also that somehow the tabs vs. newline convention has gotten mixed in
at some point.

We can start by removing all hard tabs in favor of whitespace by
running Emacs from an org file over the source units.