Start out with the actual Class in print-object on java:java-object
This prevents jclass-superclass from failing on names (keywords) of
classes generated by jnew-runtime-class (it fails because they are
created in a different ClassLoader and it can't find them). By passing
an actual Class it doesn't have to do a lookup by name. The fail
occured on the first iteration of the loop in
print-java-object-on-class.

NB: Yes, jcall works on keywords. So this is working code:
  (jcall "getName" :|java.util.List|)
  (jclass-superclass :|java.util.HashMap|)

The inheritance mechanism of print-java-object-by-class is unchanged,
because it internally converts the classes found by jclass-superclass
into a keyword in order to find-method. Before this change, it also
worked because of the above fact that jcall getName works on keywords.

For jnew-runtime-class hierarchies this is fine as well, since the
programmers specialize print-object-by-class themselves on any keywords
they want, so they are responsible for any bugs in that. I.E ABCL's
catch-all print-java-object-by-class doesn't ever call
jclass-superclass on a keyword because the `super` loop variable there
is always java.lang.Class or nil (that's what jclass-superclass
returns). Before this patch, it was a keyword on the first iteration,
then a java.lang.Class in the rest of the iterations.

Passing a Class is better than passing in a keyword and making it look
up the Class by name (an information loss at the start of the function
- it goes back to looping on actual Classes after the first iteration
anyway) because it fixes the case of printing instances of
jnew-runtime-class generated classes without having to play around with
classloaders.