ci: configure for openjdk19 and beyond
It should be enough to use just the source option, and assume we will
be targeting running on the JDK we are using for compilation.

TODO: for the 'abcl.release' Ant build, we will want to be more
selective about target.

TODO:  add a build.xml target that just passes the source option. use
it for the default.