java: provide api for getting and setting classloader
The WITH-CLASSLOADER macro will set the context classloader to that of
the provided CLASSLOADER, or use the default ABCL classloader if
specfied with no argument.

The CLASSLOADER function returns the default classloader or that
associated with a specified JAVA-OBJECT, or the default ABCL
classloader if no JAVA-OBJECT is given

The CONTEXT-CLASSLOADER function returns the context classloader
associated with a JAVA-THREAD, or the currently executing thread if no
JAVA-THREAD is specified.  The CONTEXT-CLASSLOADER is a SETF'able
place.