ci: update openjdk versions
Deprecate openjdk19 in favor of openjdk20 as it is the current
development release.

Have <file:ci/set-jdk.bash> only use its arguments to set jdk
version, while attempting to be more verbose and guard against failure
to interpolate these values as we go.

Usage:

    set-jdk OPENJDK_VERSION ABCL_ROOT

where

OPENJDK_VERSION is of the form 'openjdk8', 'openjdk19', etc.

ABCL_ROOT is the root of the ABCL source tree