Fix for SUBTYPEP on selected types
Even though as @alejandrozf points out in
<https://github.com/armedbear/abcl/issues/433#issuecomment-1024998558>
that the original behavior is conforming, it is more helpful for the
user to fix this.

This is a rather hacky fix as we "merely" special-case the problems,
meaning there are probably many more edge cases.  Our implementation
of SUBTYPEP could use more sophistication, as the proper relationships
are correctly specified in our built-in class precedence lists from I
can see.  But SUBTYPEP is seemingly too low-level to use these
relations easily, as they are expressed in the Java-side code.

See <https://github.com/armedbear/abcl/issues/433>.