ci: update openjdk installations to latest
Stop depending on the GitHub installed CI Java.  It used to be that
with GitHub actions/setup-java@v1 we could Ant installed but use jenv
to use our "own" openjdk to run it.  Something changed, debugging the
GitHub CI via multiple pull requests is tedious, so just try to use
GitHub for any dependencies beyond bash.

Robustify syntax for specifying operating system and architecture to
install-openjdk.