Properly declare global-scope script context bindings as special.
With the declaration in the inner engine-scope `let`, _references to_ the
bindings for global-scope vars are declared to be special, but the bindings
themselves aren't; the global vars are bound lexically, which causes two
problems:
1. the vars are unbound in the script, and can't be accessed; and
2. the put-bindings jcall at the end of script execution fails because
   the variable is unbound.

The latter occurs because the only binding of that variable is lexical, but
the special-var declaration in the inner `let` has declared that all
references to that var are to a dynamic binding, not a lexical binding.
There is no dynamic binding, so the var is unbound and passing it as an arg
signals an unbound-var error.

Solve both problems by declaring the global-scope vars as special in the
`let` that creates the bindings, like we do for engine-scope vars.