ci: unpin ansi-test to master
Improve runnabilty in more places and readibility of progress.

The 'ansi-test' repository doesn't see a lot of updates, but using the
latest published 'master' branch will enable more rapid debugging:
just beware in comparing test results of CI runners that the version
of the ANSI-TEST suite be used may change without warning.  In that
case, we can repin things.  -- mgmt.